+++ pacs-server/src/application/use_cases/mod.rs
@@ -4,6 +4,8 @@ pub mod project_use_case;
 pub mod permission_use_case;
 pub mod access_control_use_case;
 pub mod annotation_use_case;
+pub mod mask_group_use_case;
+pub mod mask_use_case;
 
 pub use auth_use_case::AuthUseCase;
 pub use user_use_case::UserUseCase;
@@ -11,3 +13,5 @@ pub use project_use_case::ProjectUseCase;
 pub use permission_use_case::PermissionUseCase;
 pub use access_control_use_case::AccessControlUseCase;
 pub use annotation_use_case::AnnotationUseCase;
+pub use mask_group_use_case::MaskGroupUseCase;
+pub use mask_use_case::MaskUseCase;